% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.manhattan.R
\name{plot.manhattan}
\alias{plot.manhattan}
\title{Create a nice looking manhattan plot of significantly up- and downregulated genes.}
\usage{
plot.manhattan(raw.count.matrix, design, name = "", file.loc = "",
  title.name = "", count.mean.cutoff = 10, logFC.cutoff = 0.8,
  Pvalue.cutoff = 0.05, write.all.genes = T, reduce.point.size.by = 6,
  dpi = 300, width = 297, height = 210, units = "mm", open.pdf = F,
  open.html = F, which.attribute = 1, reverseGroup = F,
  highligt.top.genes = F, nr_top_genes = 10, nr_bottom_genes = 10,
  which.biotypes = "", inverse.biotypes = F, order.by.pvalue = F,
  write.cpm = F)
}
\arguments{
\item{raw.count.matrix}{Samples in columns. Genes in rows.}

\item{design}{Design data.frame consists with row.names = samples, and columns for each clinical attribute. Each column in the design can only have a maximum of two levels. Any NA attributes will be removed.}

\item{name}{= "" If you would like to specify the output names of each file manually. Defaults to column name of clinical attribute.}

\item{title.name}{Title name in plot. Defaults to column name of clinical attribute.}

\item{count.mean.cutoff}{Cutoff before running edgeR analysis (default = 10)}

\item{logFC.cutoff}{Cutoff before correcting p values (default = 0.8)}

\item{Pvalue.cutoff}{Plot only genes that satisfy this criteria ( default = 0.05)}

\item{write.all.genes}{Wheter to write all genes in edgeR output, or to only write those below Pvalue.cutoff (default = F)}

\item{reduce.point.size.by}{Useful when point sizes are too big (or too low)}

\item{dpi}{Pixels per inch in pdf created (default = 300)}

\item{width}{Width of pdf (default = 297)}

\item{height}{Height of pdf (default = 210)}

\item{units}{Units for width and height (default = "mm")}

\item{open.pdf}{Will open the manhattan plot after the function has completed}

\item{open.html}{Will open the manhattan plot after the function has completed}

\item{reverseGroup}{If the comparison made by edgeR is not what we wanted, e.g. we would like AMPLIFIED / NON-AMPLIFIED instead of NON-AMPLIFIED / AMPLIFIED (default = FALSE)}

\item{highligt.top.genes}{Highlight top differentially expressed genes in plot.}

\item{nr_top_genes}{How many genes with positive fold change values should be labelled. Default = 10.}

\item{nr_bottom_genes}{How many genes with negative fold change values should be labelled. Default = 10.}

\item{which.biotypes}{= "protein_coding"   ## if interested in comparing only protein_coding genes}

\item{inverse.biotypes}{If interested in only long non-coding RNAs set which.biotypes = "protein_coding" and inverse.biotypes = T. Default = FALSE.}

\item{order.by.pvalue}{Order and plot by pvalue rather than FDR (default = FALSE)}

\item{write.cpm}{Write count matrix to output folder (default = FALSE)}

\item{which.biotypes}{If we would like to subset on a particular biotype. E.g. "protein_coding". Default = ""}
}
\examples{
See https://github.com/utnesp/R-manhattan.plot
}
